# Brian's Code/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KeepTally inventory tracking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Item {
  id: number;
  name: string;
  category: string;
  quantity: number;
  parLevel: number;
  location: string;
  barcode?: string | null;
  lastUpdated: string;
  createdAt: string;
}

export interface CreateItemBody {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  quantity: number;
  /** @minimum 0 */
  parLevel: number;
  /** @minLength 1 */
  location: string;
  barcode?: string | null;
}

export interface UpdateItemBody {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  category?: string;
  /** @minimum 0 */
  quantity?: number;
  /** @minimum 0 */
  parLevel?: number;
  /** @minLength 1 */
  location?: string;
  barcode?: string | null;
}

export interface HistoryEntry {
  id: number;
  /** @nullable */
  itemId?: number | null;
  itemName: string;
  /** create | update | delete | adjust | command */
  action: string;
  /**
   * Field that changed (e.g. quantity, name, location)
   * @nullable
   */
  field?: string | null;
  /** @nullable */
  previousValue?: string | null;
  /** @nullable */
  newValue?: string | null;
  /** @nullable */
  note?: string | null;
  /** ui | command */
  source: string;
  createdAt: string;
}

export interface CommandBody {
  /** @minLength 1 */
  text: string;
}

export interface CommandAction {
  /** set | add | reduce | create | delete */
  kind: string;
  /** @nullable */
  itemId?: number | null;
  itemName: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  previousQuantity?: number | null;
  /** @nullable */
  newQuantity?: number | null;
  /** @nullable */
  delta?: number | null;
}

export interface CommandResult {
  success: boolean;
  message: string;
  action?: CommandAction | null;
}

export interface RestockEntry {
  item: Item;
  unitsNeeded: number;
}

export interface RestockList {
  generatedAt: string;
  totalItems: number;
  totalUnitsNeeded: number;
  entries: RestockEntry[];
}

export type DashboardSummaryByCategoryItem = {
  category: string;
  itemCount: number;
  totalUnits: number;
};

export type DashboardSummaryByLocationItem = {
  location: string;
  itemCount: number;
  totalUnits: number;
};

export interface DashboardSummary {
  totalItems: number;
  totalUnits: number;
  belowParCount: number;
  outOfStockCount: number;
  locationsCount: number;
  categoriesCount: number;
  belowPar: Item[];
  recentChanges: HistoryEntry[];
  byCategory: DashboardSummaryByCategoryItem[];
  byLocation: DashboardSummaryByLocationItem[];
}

export type ListItemsParams = {
  /**
   * Filter by location name
   */
  location?: string;
};

export type ListHistoryParams = {
  /**
   * Filter by item id
   */
  itemId?: number;
  limit?: number;
};

export type GetRestockListParams = {
  /**
   * Filter by location name
   */
  location?: string;
};

export type GetDashboardSummaryParams = {
  /**
   * Filter summary by location name
   */
  location?: string;
};