# Brian's Code/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KeepTally inventory tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: High-level inventory summary
      parameters:
        - in: query
          name: location
          required: false
          schema:
            type: string
          description: Filter summary by location name
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
components:
  schemas:
...
    Item:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        quantity:
          type: integer
        parLevel:
          type: integer
        location:
          type: string
        lastUpdated:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - quantity
        - parLevel
        - location
        - lastUpdated
        - createdAt
...
    HistoryEntry:
      type: object
      properties:
        id:
          type: integer
        itemId:
          type: ["integer", "null"]
        itemName:
          type: string
        action:
          type: string
          description: "create | update | delete | adjust | command"
        field:
          type: ["string", "null"]
          description: Field that changed (e.g. quantity, name, location)
        previousValue:
          type: ["string", "null"]
        newValue:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        source:
          type: string
          description: "ui | command"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - itemName
        - action
        - source
        - createdAt
...
    DashboardSummary:
      type: object
      properties:
        totalItems:
          type: integer
        totalUnits:
          type: integer
        belowParCount:
          type: integer
        outOfStockCount:
          type: integer
        locationsCount:
          type: integer
        categoriesCount:
          type: integer
        belowPar:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        recentChanges:
          type: array
          items:
            $ref: "#/components/schemas/HistoryEntry"
        byCategory:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              itemCount:
                type: integer
              totalUnits:
                type: integer
            required:
              - category
              - itemCount
              - totalUnits
        byLocation:
          type: array
          items:
            type: object
            properties:
              location:
                type: string
              itemCount:
                type: integer
              totalUnits:
                type: integer
            required:
              - location
              - itemCount
              - totalUnits
      required:
        - totalItems
        - totalUnits
        - belowParCount
        - outOfStockCount
        - locationsCount
        - categoriesCount
        - belowPar
        - recentChanges
        - byCategory
        - byLocation